Añadido deltaTime

+++ Practica1/Engine/src/main/java/es/fdi/ucm/gdv/vdism/maranwi/engine/Application.java
@@ -5,7 +5,7 @@ public interface Application {
     public boolean onExit();
     public void onRelease();
     public void onInput(Input input);
-    public void onUpdate(float deltaTime);
+    public void onUpdate(double deltaTime);
     public void onRender(Graphics graphics);
     public void setApplicationZone(double width, double height, double windowsWidth, double windowsHeigth);
     public int getLogicWidth();
