+++ Practica1/Engine/src/main/java/es/fdi/ucm/gdv/vdism/maranwi/engine/GameState.java
@@ -5,7 +5,7 @@ import es.fdi.ucm.gdv.vdism.maranwi.engine.Graphics;
 public interface GameState {
     public void start(Graphics g);
     public void render(Graphics g);
-    public void update(float deltaTime);
+    public void update(double deltaTime);
     public void identifyEvent(int x, int y);
     public void setMainApplicaton(Application a);
 }
